backend/requirements.txt
fastapi
uvicorn
python-multipart
firebase_admin
torch
safetensors
pytesseract 
motor 
pymongo 
python-jose 
pdfminer.six
reportlab
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
pymupdf 
flask 
pdf2image 
PyPDF2
opentelemetry-instrumentation-logging
bs4
"pydantic[email]"
"uvicorn[standard]"